fine-tuning of build system, now it works even easier

+++ 3rdparty/libraries/curl/build_x86.sh
@@ -4,6 +4,12 @@
 #
 echo "Executing x86 build script"
 
+# Check if we already built for this target, run clean.sh to reset everything.
+if [ -d x86/lib/ ]; then
+	echo "Target already built, done executing x86 build script"
+	exit 0; 
+fi
+
 DIR_PREFIX=$PWD/x86
 
 rm -rf curl-7.18.1
